NeuralStyleTransfer/requirements.txt
scipy==1.1.0
numpy==1.15.4
tensorflow==1.12.0
Pillow==5.3.0
